[no-issue] fix: bugRender 수정
